refactors handlers, middleware pkg
