+++ cmd/server/middleware/buffering_writer.go
@@ -1,4 +1,4 @@
-package main
+package middleware
 
 import (
 	"bufio"
