+++ cmd/server/middleware/openapi_response_log.go
@@ -1,4 +1,4 @@
-package main
+package middleware
 
 import (
 	"bytes"
